Added IsHalfEmpty and Shift functionality
the simple queue has a drawback that after a certain no. of enqueue and dequeue operations, the initial parts of the queue cannot be filled again before emptying the whole queue. the IsHalfEmpty function checks if the queue is half empty and if true, shifts all elements to the start of the queue